Once upon a time, in a land far, far away, I wrote this first implementation of Atlantis. It uses a then-brand-new HTML5 canvas element to draw the board, and features real-time play between players through AJAX requests to a PHP server that uses (drumroll...) named pipes for inter-request coordination.
The code is added to this repository mostly for archeological interest, as we intend to rewrite most of the game from scratch. However, in case anybody ever feels like re-installing this code somewhere, take note that you'll need to edit 'config.php' and 'globals.js', as well as provide writable subdirectories named 'games' and 'pipes'.