Implement tentative moves, turn-based play, fix movement rules, and add unit testing.
We alter the database schema to support the concept of turns and board positions within those turns (as opposed to just a 'moves' counter). We make the web client commit changes in that style; it now knows about when its turn is finished and signals that. We implement UI limitations to let only the player whose turn it is do moves.

There were a number of important missing movement constraints, such as not moving from/onto/over dead/growing points. We add some graceful styling around full annihilation in the tentative moves UI.

We add unit testing for data structures and movement rules, since they can get complex.